+++ Assets/Scripts/Mayotech/Resource/OnResourceChangedEvent.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 namespace Mayotech.Resources
 {
     [CreateAssetMenu(menuName = "GameEvent/OnResourceChangedEvent")]
-    public class OnResourceChangedEvent : GameEvent<LocalResource, int>{}
+    public class OnResourceChangedEvent : GameEvent<LocalResource, long>{}
 }
